ref: rename database to models #6
